requirements.txt
langchain_ollama
langchain_community
langchain_chroma
pypdf
